Adds animation on resize for share profile dialog (#65)
# Description & Technical Solution
- Adds `react-use-measure` package that calculates height of a component
- Implements framer motion to animate height changes
- Changes `customDialogClose: string` prop to `hideDefaultCloseButton:
boolean` in shadcn `dialog` component for better readability.

** Opted not to add the horizontal animations because I felt that it
would be too much and very distracting. Felt that opacity and height
resize animation is clean and sleek. Open to discussion tho!

If it were something like this:
https://ui.lukacho.com/components/dropdown-menu where a user has a sense
of direction because everything is displayed, then i think a horizontal
slide makes sense. I digress, but these are just my thoughts.

# Screenshots


https://github.com/user-attachments/assets/3d3b3ebc-c809-4345-9753-ae5910eb8a60