Fixes profile updates not taking effect immediately SRBT-169 (#69)
# Description & Technical Solution
- Implements RQ cache invalidation to refetch user data at the key of
`freelancer`
- Also adds comments to `LocationPredictionsList`
- Also fixes some linting warnings

** I think the method of presenting the data may need some refactoring
because we are making an unnecessary network request. In the grand
scheme of things, this is a very small improvement and might be
considered negligible, but I think there is room for consistency.

As it stands, if you are on your own profile or not, we are making a
network request to fetch the data of the user in the URL params. I think
that this should happen only if you are visiting someone else's profile
because we have all the data we need already in `useAuth`.

An example of inconsistency is that for the user profile data on the
widgets page, we are rendering data fetched with RQ, but in the
`header.tsx` component in the top right, we are using data from the
`useAuth` hook. (This is when a user is looking at their own profile). I
think that RQ is good and should be used, but only in the scenario where
a user that is signed in is looking at someone else's profile.

# Video fixing profile update bug (unrelated to the consistency note
above)



https://github.com/user-attachments/assets/5a25e787-9175-467f-8667-5bc12c1d0717