[SRBT-273] Gigs polish (#110)
# [SRBT-273] Gigs polish 

**Changes**

- Updates UI to better match Figma designs. All updated components have
a `.stories` file that you can access and see locally.
- Unused buttons in the chat bar have been removed
- `gigs-dialog.tsx` overflow issue has been resolved (as of the time of
writing this, there may be a tweak that needs to be made to the
DialogContent component)
- Ended up opting to not migrate to React SDK due to the fact that we
have more flexibility when we use the JS SDK and have to eventually
comeback to the JS SDK. Ex: I was having issues with setting up a
GroupChannelHandler because the types weren't updated and returning
properly. The solutions online were to use JS SDK to initialize a
handler and then pass it to React SDK handler. Also the docs are not
very good :(

# Screenshots

![Screenshot 2024-10-10 at 2 57
24 PM](https://github.com/user-attachments/assets/4f7a2c9f-b006-44f6-9c9a-922c58a1c6ff)

![Screenshot 2024-10-10 at 2 57
40 PM](https://github.com/user-attachments/assets/8df24f43-8cd1-4f93-a2cc-d7231bf66aa9)


https://github.com/user-attachments/assets/88b96220-0036-4325-8f7a-511d49e4e937