[SRBT-375] Add loading state if widget is being modified (#152)
# [SRBT-375] Add loading state if widget is being modified

Due to the changes explained below, this PR also addresses
[SRBT-345](https://linear.app/mysorbet/issue/SRBT-345/look-into-dual-toast-error-for-section-titles).

**Changes**

* Refactored the loading states of the profile page to instead only put
a widget that's being directly modified (i.e. new image upload,
restoring image, etc.) into a loading state.
* Notably, I refactored several of the `onClick` functions in
`widget.tsx` to resolve some of the deeper issues with race conditions
happening on the profile page, like so:
  * old way: `onClick={(e) => {
                  handleRemove(identifier);
                }}`
  * new way: `onMouseDown={(e) => {
                  e.stopPropagation();
                  handleRemove(identifier);
                }}`
* When `onClick` is used, it's technically consider a "dragging" motion
by
[react-grid-layout](https://github.com/react-grid-layout/react-grid-layout/issues/166),
and thus `handleLayout` was getting triggered and subsequently
`persistWidgetsLayoutOnChange` due to the change to `layout`. So there
was two outgoing calls to the bulk-update endpoint in the following
order: one where it thought the widget is just getting moved in the
grid, and one where it was considered deleted (which is why it was
inconsistent on when failures happen).

# Screenshots

Screenshots/videos of the changes made (if any, otherwise delete this
section)

---------

Co-authored-by: Audrey Ostrom <audreyostrom64@gmail.com>