[SRBT-355] Instagram widget image replacement and restoring (#154)
# [SRBT-355] Instagram widget image replacement and restoring

**Changes**

- Users can use the instagram profile's most recent posts, or they have
the option of using a cover image. The precedent of which gets
displaying is instagram posts (if they exist) -> cover image -> whatever
appropriate placeholder text for when there's neither.
- Simplified `restoreWidgetImageAsync` to not need a content parameter
to reduce payload size. Wasn't even needed for the corresponding backend
function to work anyways.
- For handling image restorations for instagram widgets, a new endpoint
was made and that gets called by `restoreInstagramWidgetAsync` instead
of `restoreWidgetImageAsync`. For context, each individual instagram
image is ~500 kb, which means the flow of fetching an image and then
updating the widget with said image doesn't work because because that
would make the payload of that second request too big.
- Change `onClick` to `onMouseDown` in `modify-image-controls.tsx` for
reasons described in this previous
[PR](https://github.com/MySorbet/sorbet-app/pull/152)

# Screenshots


https://github.com/user-attachments/assets/da177173-b25a-47c1-bf80-f018491c78f9

Ignore the flicker when I replace the image around ~0:30. That's due to
me forgetting to refactor the `onClick` to `onMouseDown` (it's the
widget moving positions because persistLayout got called twice). That's
fixed now.

---------

Co-authored-by: Audrey Ostrom <audreyostrom64@gmail.com>